refactor(mcp-manager): 拆解 makeRoutes 与 apply 装配骨架，圈复杂度 84/78 降至 5/14 (#592)
- makeRoutes（comp 84→5）：9 个端点 handler 拆离至 routes-controllers.ts
  独立控制器工厂（buildXxxRoute(manager, helpers)），connect/disconnect/
  reconnect 三兄弟路由收敛为 buildNameActionRoute 同构工厂；routes.ts 退化为
  ROUTES 常量 + 装配 + SSE/health 工厂；共享 helpers（handleError/scopeParam/
  maybeSession）显式参数化，queryParam 提升至 routes-helpers.ts
- apply（comp 78→14）：配置解析与 settings 接线拆至 apply-config.ts
  （resolveApplyOptions/resolveMiddlewareMode/installConfigSettings/
  injectSettingsSink，含 #125 this 绑定语义原样保留）；provide 服务对象拆至
  apply-services.ts（provideMcpManagerService，8 方法契约面逐字保留）；
  fs.watch 防重入/热切换/catalog 注入/路由广播拆至 apply-runtime.ts；
  MCP_GUIDANCE 拆至 apply-guidance.ts
- service-contract.test.ts provide 源文本扫描源同步迁移至 apply-services.ts
  （fail-loud 语义不变）；全部既有 smoke/unit 契约断言零修改通过
- CRAP --diff 增量门禁 0 劣化；五连门禁全绿

Refs #592
