+++ packages/dsh-mcp-manager/src/routes.ts
@@ -1,26 +1,40 @@
 /**
- * dsh-mcp-manager — HTTP 路由（依赖 import 模块与仓库共享层）。
+ * dsh-mcp-manager — HTTP 路由装配层（依赖 import 模块与仓库共享层）。
  *
  * /api/dsh-mcp/* 路由（loopback-only）供 web GUI 分级展示、快速接入、
  * 导入 mcpServers JSON；events 为 SSE 状态推送通道。
  * 例外：/api/dsh-mcp/config 是只读 UI 配置接口，允许非 loopback 访问，
  * 便于远程页面读取 position/offset 等非敏感展示配置。
  * 由 lib/index.js 组合根 re-export（ROUTES / makeRoutes）。
+ *
+ * #592 阶段二 Batch A 消峰：各端点 handler 已拆离至 routes-controllers.ts
+ * 的独立控制器工厂（原 makeRoutes 单函数圈复杂度 84 → 装配后 <= 15）；
+ * 本文件只保留 ROUTES 常量、装配函数与 SSE/health 工厂（SSE/health 原本
+ * 即独立工厂形态，复杂度低，随装配层同文件保留）。
  */
 
 // 辅助函数统一来自仓库共享层（loopback 围栏 / writeJson / readJsonBody / sseData）。
-import { writeJson, readJsonBody, sseData, guardLoopbackMethod } from "../../../shared/host-utils.js";
+import { writeJson, sseData, guardLoopbackMethod } from "../../../shared/host-utils.js";
 import { createSseHub } from "../../../shared/sse-hub.js";
 import type { SseHub } from "../../../shared/sse-hub.js";
-import { parseClaudeJson } from "./import.ts";
-import { SCOPE_PROJECT, normalizeScope } from "./scope.ts";
-import { parseFullServerName, MIDDLEWARE_GLOBAL_ROOT, normalizeToolName } from "./middleware-utils.ts";
-import { normalizeMiddlewareMode } from "./middleware-const.ts";
 import type { ServerConfig } from "./types.ts";
 import type { ClientUiConfig } from "./index.ts";
 import type { WebRoute } from "@deepseek-ai/dsh-host-webserver";
 import type { ServerResponse } from "node:http";
 import type { McpStore } from "./store.ts";
+import {
+  buildConfigRoute,
+  buildServersRoute,
+  buildSessionRoute,
+  buildResumeRoute,
+  buildConnectRoute,
+  buildDisconnectRoute,
+  buildReconnectRoute,
+  buildImportJsonRoute,
+  buildToolDisableRoute,
+  type RouteHelpers,
+} from "./routes-controllers.ts";
+import { queryParam } from "./routes-helpers.ts";
 
 /** routes 使用的 McpManager 最小面（避免 index↔routes 循环 import）。 */
 export interface RoutesManager {
@@ -79,12 +93,7 @@ export const ROUTES = {
   toolDisable: "/api/dsh-mcp/tool-disable",
 };
 
-const MAX_JSON_BODY_BYTES = 2 * 1024 * 1024;
-
-function queryParam(url: URL, name: string): string | undefined {
-  const value = url.searchParams.get(name);
-  return value === null ? undefined : value;
-}
+export { queryParam };
 
 /** 组装 /api/dsh-mcp/* 路由。cwd 参数仅用于兼容旧调用（不再被路由使用）。 */
 export function makeRoutes(manager: RoutesManager, cwd = process.cwd()): WebRoute[] {
@@ -92,316 +101,24 @@ export function makeRoutes(manager: RoutesManager, cwd = process.cwd()): WebRout
     writeJson(res, 400, { error: error instanceof Error ? error.message : String(error) });
   };
   /** 解析 scope 查询参数（缺省 global）。 */
-  const scopeParam = (url: URL) => (normalizeScope(queryParam(url, "scope") ?? ""));
+  const scopeParam = (url: URL) => queryParam(url, "scope") ?? "";
   /** 若带 cwd 参数则先切换会话（跟随会话的项目级 MCP）。 */
   const maybeSession = async (url: URL): Promise<void> => {
     const cwdParam = queryParam(url, "cwd");
     if (cwdParam !== undefined && cwdParam !== "") await manager.setSession(cwdParam);
   };
+  const helpers: RouteHelpers = { handleError, scopeParam, maybeSession };
 
   return [
-    {
-      kind: "exact",
-      path: ROUTES.config,
-      handler: async (req, res) => {
-        // GET：只读 UI 配置 + 中间层模式（允许非 loopback，供远程页面读取非敏感的展示配置）。
-        if (req.method === "GET") {
-          try {
-            writeJson(res, 200, { ...manager.uiConfig(), middleware: manager.middlewareMode ?? "off" });
-          } catch (error) {
-            handleError(res, error);
-          }
-          return;
-        }
-        // POST：写入浮窗 UI 配置（position / offset），或热切换中间层模式
-        // （middleware: off/project/all）。写操作只对 loopback 开放；
-        // 经设置命名空间落盘（Config.ui），触发 scope.watch → onChange → SSE 广播一帧，
-        // 客户端收到后重新 GET /config 就地更新浮窗位置，无需重启/轮询。
-        if (req.method === "POST") {
-          if (!guardLoopbackMethod(req, res, ["POST"])) return;
-          let body: unknown;
-          try {
-            body = await readJsonBody(req);
-          } catch {
-            writeJson(res, 400, { error: "invalid JSON body" });
-            return;
-          }
-          if (typeof body !== "object" || body === null) {
-            writeJson(res, 400, { error: "invalid JSON body" });
-            return;
-          }
-          try {
-            const rec = body as Record<string, unknown>;
-            if (typeof rec.middleware === "string") {
-              // 中间层模式热切换：先热生效（当前进程立即切换），再落盘（重启保留）。
-              if (typeof manager.setMiddlewareMode === "function") {
-                await manager.setMiddlewareMode(rec.middleware);
-              }
-              if (typeof manager.uiUpdate === "function") {
-                await manager.uiUpdate({ middleware: normalizeMiddlewareMode(rec.middleware) });
-              }
-              writeJson(res, 200, { ...manager.uiConfig(), middleware: manager.middlewareMode ?? "off" });
-              return;
-            }
-            writeJson(res, 200, await manager.updateUiConfig(body));
-          } catch (error) {
-            handleError(res, error);
-          }
-          return;
-        }
-        // 端点级方法分流先于 loopback 的刻意例外（#473 R2 结构 β）：本 else 对
-        // 白名单外方法（PUT/DELETE/OPTIONS 等）直接 405、不查 loopback——非
-        // loopback+PUT 返回 405 而非 403 是契约行为，禁止误套守卫（会漂移为 403）。
-        writeJson(res, 405, { error: `method not allowed: ${req.method}` });
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.servers,
-      handler: async (req, res) => {
-        const url = new URL(req.url ?? "/", "http://localhost");
-        const method = req.method ?? "GET";
-        if (!guardLoopbackMethod(req, res, ["GET", "POST", "PATCH", "DELETE"])) return;
-        if (method === "GET") {
-          try {
-            // 变更点驱动（#111/#228）：GET /servers 是纯读快照，零副作用——
-            // 磁盘变更由 fs.watch 事件驱动 reconcile，状态从内存 store 读。
-            // 历史 refreshFromDisk 副作用（mtime 轮询 + reconcile）已移除。
-            // cwd 查询参数不处理（#324）：GET 带 cwd 曾触发 setSession → emitStatus
-            // → SSE 广播 → 客户端再 GET 的自激循环（多页面互踩）。会话切换只走
-            // POST /api/dsh-mcp/session；此处忽略 cwd 以恢复纯读语义。
-            writeJson(res, 200, manager.summary());
-          } catch (error) {
-            handleError(res, error);
-          }
-          return;
-        }
-        if (method === "POST") {
-          const body = await readJsonBody(req);
-          if (body === undefined) {
-            writeJson(res, 400, { error: "invalid JSON body" });
-            return;
-          }
-          try {
-            const rec = body as Record<string, unknown>;
-            const scope = normalizeScope(rec.scope as string);
-            if (typeof rec.cwd === "string" && rec.cwd !== "") await manager.setSession(rec.cwd);
-            const server = await manager.add(rec, scope);
-            writeJson(res, 201, { server, summary: manager.summary() });
-          } catch (error) {
-            handleError(res, error);
-          }
-          return;
-        }
-        if (method === "PATCH" || method === "DELETE") {
-          const name = queryParam(url, "name");
-          if (name === undefined || name === "") {
-            writeJson(res, 400, { error: "name query parameter is required" });
-            return;
-          }
-          try {
-            await maybeSession(url);
-            const scope = scopeParam(url);
-            if (method === "DELETE") {
-              await manager.remove(name, scope);
-              writeJson(res, 200, { ok: true, summary: manager.summary() });
-            } else {
-              const body = await readJsonBody(req);
-              if (body === undefined) {
-                writeJson(res, 400, { error: "invalid JSON body" });
-                return;
-              }
-              const server = await manager.update(name, body as Record<string, unknown>, scope);
-              writeJson(res, 200, { server, summary: manager.summary() });
-            }
-          } catch (error) {
-            handleError(res, error);
-          }
-          return;
-        }
-        writeJson(res, 405, { error: `method not allowed: ${method}` });
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.session,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["POST"])) return;
-        const body = await readJsonBody(req);
-        if (body === undefined || typeof (body as Record<string, unknown>).cwd !== "string") {
-          writeJson(res, 400, { error: "body must include a cwd string" });
-          return;
-        }
-        try {
-          await manager.setSession((body as Record<string, unknown>).cwd as string);
-          writeJson(res, 200, { ok: true, summary: manager.summary() });
-        } catch (error) {
-          handleError(res, error);
-        }
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.resume,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["POST"])) return;
-        try {
-          if (typeof manager.resumeReconnect !== "function") throw new Error("resumeReconnect unavailable");
-          await manager.resumeReconnect();
-          writeJson(res, 200, { ok: true, summary: manager.summary() });
-        } catch (error) {
-          handleError(res, error);
-        }
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.connect,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["POST"])) return;
-        const url = new URL(req.url ?? "/", "http://localhost");
-        const name = queryParam(url, "name");
-        if (name === undefined || name === "") {
-          writeJson(res, 400, { error: "name query parameter is required" });
-          return;
-        }
-        try {
-          await maybeSession(url);
-          await manager.connect(name, scopeParam(url));
-          writeJson(res, 200, { ok: true, summary: manager.summary() });
-        } catch (error) {
-          handleError(res, error);
-        }
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.disconnect,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["POST"])) return;
-        const url = new URL(req.url ?? "/", "http://localhost");
-        const name = queryParam(url, "name");
-        if (name === undefined || name === "") {
-          writeJson(res, 400, { error: "name query parameter is required" });
-          return;
-        }
-        try {
-          await maybeSession(url);
-          await manager.disconnect(name, scopeParam(url));
-          writeJson(res, 200, { ok: true, summary: manager.summary() });
-        } catch (error) {
-          handleError(res, error);
-        }
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.reconnect,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["POST"])) return;
-        const url = new URL(req.url ?? "/", "http://localhost");
-        const name = queryParam(url, "name");
-        if (name === undefined || name === "") {
-          writeJson(res, 400, { error: "name query parameter is required" });
-          return;
-        }
-        try {
-          await maybeSession(url);
-          await manager.reconnect(name, scopeParam(url));
-          writeJson(res, 200, { ok: true, summary: manager.summary() });
-        } catch (error) {
-          handleError(res, error);
-        }
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.importJson,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["POST"])) return;
-        const body = await readJsonBody(req);
-        if (body === undefined || typeof (body as Record<string, unknown>).json !== "string") {
-          writeJson(res, 400, { error: "body must include a json string" });
-          return;
-        }
-        const imported: string[] = [];
-        const skipped: string[] = [];
-        try {
-          const rec = body as Record<string, unknown>;
-          const scope = normalizeScope(rec.scope as string);
-          if (typeof rec.cwd === "string" && rec.cwd !== "") await manager.setSession(rec.cwd);
-          const store = scope === SCOPE_PROJECT ? await manager.projectStoreOrThrow() : manager.store;
-          const servers = parseClaudeJson(rec.json as string);
-          for (const server of servers) {
-            if (store.find(server.name) !== undefined) {
-              if (rec.overwrite !== true) {
-                skipped.push(server.name);
-                continue;
-              }
-              await manager.update(server.name, server, scope);
-              imported.push(server.name);
-              continue;
-            }
-            await manager.add(server, scope);
-            imported.push(server.name);
-          }
-        } catch (error) {
-          handleError(res, error);
-          return;
-        }
-        writeJson(res, 200, { imported, skipped, summary: manager.summary() });
-      },
-    },
-    {
-      kind: "exact",
-      path: ROUTES.toolDisable,
-      handler: async (req, res) => {
-        if (!guardLoopbackMethod(req, res, ["PATCH"])) return;
-        const url = new URL(req.url ?? "/", "http://localhost");
-        const body = await readJsonBody(req);
-        if (body === undefined || typeof body !== "object" || body === null) {
-          writeJson(res, 400, { error: "invalid JSON body" });
-          return;
-        }
-        const rec = body as Record<string, unknown>;
-        const server = typeof rec.server === "string" ? rec.server : "";
-        const tool = typeof rec.tool === "string" ? rec.tool : "";
-        const disabled = rec.disabled === true;
-        if (server === "" || tool === "") {
-          writeJson(res, 400, { error: "server 与 tool 均为必填" });
-          return;
-        }
-        // 路由一致性：server 全名 root 必须属于当前工作空间（或 all 模式 @global）。
-        const parsed = parseFullServerName(server);
-        if (parsed === undefined) {
-          writeJson(res, 400, { error: "server 格式非法，应为 @<root>/<server>" });
-          return;
-        }
-        const root = parsed.root === MIDDLEWARE_GLOBAL_ROOT ? parsed.root : parsed.root;
-        const cwdParam = queryParam(url, "cwd");
-        if (cwdParam !== undefined && cwdParam !== "") await manager.setSession(cwdParam);
-        const sessionRoot = manager.projectRoot;
-        const allowed = root === MIDDLEWARE_GLOBAL_ROOT || root === sessionRoot;
-        if (!allowed) {
-          writeJson(res, 400, { error: `server ${JSON.stringify(server)} 不属于当前工作空间；路由一致性校验失败（防跨空间串台）` });
-          return;
-        }
-        if (typeof manager.setToolDisabled !== "function") {
-          writeJson(res, 400, { error: "tool-disable not writable: middleware unavailable" });
-          return;
-        }
-        try {
-          // #392 遗留④：tool 参数先归一化（剥 mcp__<server>__ 前缀）再入禁用表——
-          // 旧客户端/手工 API 可能提交带前缀名，此前原样存键导致 guard 层查裸名不命中、
-          // 禁用静默无效。跨 server 前缀（剥后仍 mcp__ 开头）由 normalizeToolName 抛错。
-          const toolName = normalizeToolName(parsed.server, tool, "tool-disable");
-          await manager.setToolDisabled(root, parsed.server, toolName, disabled);
-          writeJson(res, 200, { ok: true, summary: manager.summary() });
-        } catch (error) {
-          handleError(res, error);
-        }
-      },
-    },
+    buildConfigRoute(manager, helpers),
+    buildServersRoute(manager, helpers),
+    buildSessionRoute(manager, helpers),
+    buildResumeRoute(manager, helpers),
+    buildConnectRoute(manager, helpers),
+    buildDisconnectRoute(manager, helpers),
+    buildReconnectRoute(manager, helpers),
+    buildImportJsonRoute(manager, helpers),
+    buildToolDisableRoute(manager, helpers),
   ];
 }
 
@@ -469,7 +186,7 @@ export function makeEventsRoute(manager: RoutesManager, options?: { heartbeatMs?
       const hub = (manager.sseHub ??= createSseHub({
         getMaxConnections: () => options?.maxConnections ?? 16,
         heartbeatMs: options?.heartbeatMs,
-        warn: (message) => {
+        warn: (_message) => {
           // 无 logger 注入面，静默（连接回收是自愈路径，无需告警）
         },
       }));
